requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.7
django-environ==0.11.2
django-filter==25.1
djangorestframework==3.15.2
filelock==3.15.4
identify==2.6.0
idna==3.7
nodeenv==1.9.1
oauthlib==3.2.2
platformdirs==4.2.2
pre-commit==3.7.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tzdata==2025.2
urllib3==2.2.2
virtualenv==20.26.3
